+++ src/Storage/Storage.Management/StorageAccount/AddAzureStorageAccountNetworkRule.cs
@@ -128,6 +128,11 @@ namespace Microsoft.Azure.Commands.Management.Storage
 
         public override void ExecuteCmdlet()
         {
+            if (DryRun.IsPresent && TryHandleDryRun())
+            {
+                return; // skip real execution
+            }
+
             base.ExecuteCmdlet();
 
             if (ShouldProcess(this.Name, "Add Storage Account NetworkRules"))